`decode_insn_with_objdump` with `ArchKind`
